Allow custom animations by shifting the detail into the public API
